refactor: option services 분리
